+++ frontend/app/scripts/controllers/admin/sponsors.coffee
@@ -20,19 +20,20 @@ angular.module('irishclimbingApp')
         $location.path '/'
 
     $scope.delete = (id) ->
-        $scope.deleting[id] = true
-        $scope.success = null
-        $scope.error = null
-        $http.delete(serverUrl + 'sponsors.php/' + id)
-        .success ->
-            $scope.deleting[id] = false
-            for index, sponsor of $scope.sponsors
-                if sponsor.id == id
-                    $scope.sponsors.splice(index, 1)
-            $scope.success = 'Successfully deleted sponsor'
-        .error ->
-            $scope.deleting[id] = false
-            $scope.error = 'Error deleting sponsor'
+        if confirm 'Are you sure you want to delete this sponsor?'
+            $scope.deleting[id] = true
+            $scope.success = null
+            $scope.error = null
+            $http.delete(serverUrl + 'sponsors.php/' + id)
+            .success ->
+                $scope.deleting[id] = false
+                for index, sponsor of $scope.sponsors
+                    if sponsor.id == id
+                        $scope.sponsors.splice(index, 1)
+                $scope.success = 'Successfully deleted sponsor'
+            .error ->
+                $scope.deleting[id] = false
+                $scope.error = 'Error deleting sponsor'
 
     $scope.create = ->
         $scope.edit null
@@ -41,6 +42,17 @@ angular.module('irishclimbingApp')
         modalInstance = $modal.open
             templateUrl: 'views/admin/sponsor.html'
             controller: 'AdminSponsorCtrl'
+            keyboard: false
+            backdrop: 'static'
             resolve:
                 sponsor: ->
                     angular.copy sponsor
+        .result.then (response) ->
+            if response.id
+                $scope.success = 'Successfully edited sponsor'
+                for index, sponsor of $scope.sponsors
+                    if sponsor.id == response.id
+                        $scope.sponsors[index] = response
+            else
+                $scope.success = 'Successfully created sponsor'
+                $scope.sponsors.push response
